refactor(tests): move public-API reconciler tests from inline mod to tests/reconciler_*.rs
Migrates 36 tests out of `src/core/reconciler.rs::mod tests` into 8
per-feature integration test files under `tests/`, leaving 92 white-box
tests inline with a module-level annotation listing the private items
they exercise. Pure mechanical refactor: no production code, visibility,
or test logic changes — only test relocation with adjusted `use` paths.
Shared helpers extracted to `tests/common/reconciler_fixtures.rs` since
multiple new files reuse them.

New files (each ≤ 1500-LOC budget):
- tests/reconciler_idempotency.rs (3 tests) — idempotent + unchanged path + symlink boundary
- tests/reconciler_error_display.rs (7 tests) — ReconcileError Display + CollectionDirtyStatus
- tests/reconciler_source_audit.rs (6 tests) — source-text static analysis pins
- tests/reconciler_db_only_state.rs (5 tests) — has_db_only_state public branches
- tests/reconciler_classification.rs (4 tests) — hard-delete vs quarantine
- tests/reconciler_full_hash_invariants.rs (3 tests) — zero-row aborts + unchanged-hash
- tests/reconciler_changed_hash.rs (5 tests) — hash rotation + fail-closed
- tests/reconciler_misc.rs (3 tests) — bulk commit chunks + halt + stat-diff walk
- tests/common/reconciler_fixtures.rs — shared open_test_db, insert_collection, seed helpers

Inline residue (92 tests, 3091 LOC):
private items exercised include `apply_full_hash_metadata_self_heal`,
`apply_hash_rename_matches`, `apply_native_rename_matches`, `apply_reingest`,
`apply_uuid_rename_matches`, `authorize_full_hash_reconcile`,
`build_full_hash_plan`, `canonical_body_refusal_reason`, `capture_phase1_drift`,
`classify_missing_paths`, `collection_recovery_dir`, `db_only_state_branches`,
`DbOnlyStateBranches`, `default_restore_stability_max_iters`,
`detect_duplicate_uuids_in_tree`, `fresh_collection_dirty_status`,
`hash_refusal_reason`, `infer_type_from_path`, `is_markdown_file`,
`load_collection_by_id`, `load_db_files`, `load_frontmatter_map`,
`MissingPageIdentity`, `NativeRename`, `NewTreeIdentity`, `path_to_string`,
`raw_import_invariant_result`, `reconcile_with_native_events`, `RenameMatch`,
`RenameResolution`, `resolve_rename_resolution`, `run_phase2_stability_check`,
`run_phase3_pre_destruction_fence`, `run_restore_remap_safety_pipeline_inner`,
`sentinel_count`, `stat_diff_from_walk`, `StatSnapshot`, `strip_numeric_prefix`,
`take_stat_snapshot`, `uuid_migration_preflight`, `verify_read_only_mount`,
`walk_collection`, plus private methods `FullHashReconcileMode::as_str`,
`RestoreRemapOperation::as_str`, `DriftCaptureSummary::from_stats` /
`has_material_changes` / `add_assign` / `identity`, and the private
`RenameMatchKind` enum.

Verification:
  baseline (parent a1018eb):  passed=1357 failed=0
  this commit:                passed=1357 failed=0

Both clippy channels green:
  cargo clippy --all-targets --locked -- -D warnings
  cargo clippy --all-targets --no-default-features \
    --features bundled,online-model --locked -- -D warnings

src/core/reconciler.rs:  7411 LOC -> 6220 LOC (-1191 LOC, -16%)

Part of OpenSpec change `extract-inline-tests-to-integration`.
